# price-service-rest/src/generated-src/src/main/java/org/openapitools/api/DeleteProductPriceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "DeleteProductPrice", description = "the DeleteProductPrice API")
public interface DeleteProductPriceApi {

    /**
     * DELETE /api/v1/product-price/{productId} : Delete price for one product
     *
     * @param productId Product ID (required)
     * @return Successfully delete price for product (status code 200)
     */
    @Operation(
        operationId = "deleteProductPrice",
        summary = "Delete price for one product",
        tags = { "delete-product-price" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successfully delete price for product")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/product-price/{productId}"
    )
    ResponseEntity<Void> deleteProductPrice(
        @Parameter(name = "productId", description = "Product ID", required = true) @PathVariable("productId") Integer productId
    );

}