# price-service-rest/src/generated-src/src/main/java/org/openapitools/api/ProductPriceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ProductPriceDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "ProductPrice", description = "the ProductPrice API")
public interface ProductPriceApi {

    /**
     * GET /api/v1/product-price/{productId}/{currencyCode} : Get price for one product
     *
     * @param productId Product ID (required)
     * @param currencyCode Requested currency code (required)
     * @return Get product price successful (status code 200)
     */
    @Operation(
        operationId = "getProductPrice",
        summary = "Get price for one product",
        tags = { "product-price" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Get product price successful", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductPriceDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/product-price/{productId}/{currencyCode}",
        produces = { "application/json" }
    )
    ResponseEntity<ProductPriceDto> getProductPrice(
        @Parameter(name = "productId", description = "Product ID", required = true) @PathVariable("productId") Integer productId,
        @Parameter(name = "currencyCode", description = "Requested currency code", required = true) @PathVariable("currencyCode") String currencyCode
    );

}